# static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/hosts": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get paginated host list",
                "tags": [
                    "HostApi"
                ],
                "summary": "获取主机列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "add a new host",
                "tags": [
                    "HostApi"
                ],
                "summary": "添加主机",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Host name",
                        "name": "HostName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Host IP address",
                        "name": "HostIP",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Host port number",
                        "name": "HostPort",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Host"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/hosts/assign": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "assign host to a user",
                "tags": [
                    "HostApi"
                ],
                "summary": "分配主机给用户",
                "parameters": [
                    {
                        "description": "User and host assignment information",
                        "name": "assignment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.hostAssignment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/hosts/search": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "search hosts by host name",
                "tags": [
                    "HostApi"
                ],
                "summary": "搜索主机",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Host name for filtering",
                        "name": "host",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/hosts/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get detailed information about a host",
                "tags": [
                    "HostApi"
                ],
                "summary": "获取主机详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Host ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Host"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "update host information",
                "tags": [
                    "HostApi"
                ],
                "summary": "编辑主机信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Host ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Host information to update",
                        "name": "hostUpdates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.EditHost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Host"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "delete a host by ID",
                "tags": [
                    "HostApi"
                ],
                "summary": "删除主机",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Host ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/hosts/{id}/ssh": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "redirect to SSH access path for a host",
                "tags": [
                    "HostApi"
                ],
                "summary": "SSH连接到主机",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Host ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to SSH access path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/hosts": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get user host info",
                "tags": [
                    "HostApi"
                ],
                "summary": "获取用户已绑定主机信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}/hosts/unbind": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get list of hosts not bound to a user",
                "tags": [
                    "HostApi"
                ],
                "summary": "获取未绑定的主机列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应编码：0成功 401请登录 403无权限 500错误",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "数据对象"
                },
                "msg": {
                    "description": "消息提示",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "controller.hostAssignment": {
            "type": "object",
            "required": [
                "host_id",
                "user_id"
            ],
            "properties": {
                "host_id": {
                    "description": "HostIds []int64 `json:\"host_id\" binding:\"required\"`",
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
        "model.Host": {
            "type": "object",
            "properties": {
                "created_time": {
                    "type": "string"
                },
                "hostIP": {
                    "type": "string"
                },
                "hostName": {
                    "type": "string"
                },
                "hostPassword": {
                    "type": "string"
                },
                "hostPort": {
                    "type": "integer"
                },
                "hostUser": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_time": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                }
            }
        },
        "model.Permission": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.Role": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "角色编码",
                    "type": "string"
                },
                "created_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "*gorm.Model\nId     uint   `json:\"id\"`",
                    "type": "string"
                },
                "permission": {
                    "description": "角色权限",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Permission"
                    }
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态 1-启用 2-禁用",
                    "type": "integer"
                },
                "updated_time": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "created_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "$ref": "#/definitions/model.Role"
                },
                "roleID": {
                    "type": "integer"
                },
                "updated_time": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
        "request.EditHost": {
            "type": "object",
            "properties": {
                "hostIP": {
                    "type": "string"
                },
                "hostName": {
                    "type": "string"
                },
                "hostPassword": {
                    "type": "string"
                },
                "hostPort": {
                    "type": "integer"
                },
                "hostUser": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}